requirements.txt
absl-py==2.0.0
accelerate==0.6.0
beautifulsoup4==4.12.2
biopython==1.81
boto3==1.28.38
botocore==1.31.38
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
deepspeed==0.5.3
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.12.3
fsspec==2023.6.0
goatools==1.3.1
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lit==16.0.6
lmdb==1.4.1
Markdown==3.5
MarkupSafe==2.1.3
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==9.3.0
protobuf==3.20.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
sacremoses==0.0.53
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
seqeval==1.2.2
six==1.16.0
soupsieve==2.5
statsmodels==0.14.0
sympy==1.12
tape-proteins==0.5
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==1.8
threadpoolctl==3.2.0
tokenizers==0.10.3
torch==1.9.0+cu111
torchaudio==0.9.0
torchsummary==1.5.1
torchvision==0.10.0+cu111
tqdm==4.66.1
transformers==4.12.2
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3==1.26.16
Werkzeug==3.0.0
XlsxWriter==3.1.2
zipp==3.17.0
